Don't panic while panicking in CLI's tests
Diesel CLI's tests will abort with no output if the tests for `diesel
database create` fail to drop the database. (This is actually failing
for many people because the code to drop it has diverged from CLI's
behavior, which will be fixed separately).

This change ensures that none of the `Drop` impls in our test suite will
abort the process.