Ensure MySQL schema inference loads tables when no schema given
Reminder: MySQL has no concept of schemas, and calls databases schemas
instead. All connections have access to all databases on that server
that the user has permission to access.

On PG, the default schema is always simply called `public`. For MySQL,
it'll be based on the connection URL. In theory we could actually get
this information without executing a query, but that would require
adding a method to `Connection` explicitly for this which I don't want
to do. I'm fine with just executing a query.

As an aside, I really want to find a way to require `String:
FromSql<types::Text, Self>` as a constraint of `Backend` without forcing
that to be specified *every* time I do `DB: Backend`.

I haven't written a test case for this, since I still need to figure out
what we're going to do about tests which need to mutate schema without
fucking with each other.