Detect unique constraint violations for MySQL
MySQL doesn't expose a specific API to get information about the error
other than the code, message, and SQLSTATE value. We could potentially
get the constraint name by parsing the error message, but that is
going to be fragile at best. If someone wants to rely on that fragile
method of determining the constraint name, they should do so in their
application code directly.