Add support for MySQL to Diesel CLI
MySQL is relatively similar to PG for CLI's purposes. The main
differences are:

- `information_schema.tables` aren't scoped to the current database
- The default databases that are present are `mysql`,
  `information_schema`, and sometimes `test`. I chcose the second
  because it seemed most likely to have permissions granted
- `migration redo` can't run inside of a transaction because the
  transaction is implicitly committed when the schema is modified.
  `COMMIT` when not inside of a transaction is not an error, but
  wrapping the outer calls in a transaction cause us to use savepoints
  which error when we try to release them.

Running the tests for CLI requires GRANT ALL ON `diesel_%`.* TO
''@'localhost'. I need to see if I can make the URLs we use more
configurable.

The way we check to run a transaction in `migration_redo` is a little
hacky, but was overall less code than restructuring it to pass the enum
through further.