Clean up #667
There were a ton of style and whitespace issues. I've resolved as many
as I could. The implementation was also a bit all over the place and the
intent of the code was extremely unclear to me. I've tried to better
break things out into blocks based on the logical segments of the impl,
and what I think will need to be reused for moving future macros over to
pure procedural macro style.

I'd actually strongly prefer to have `row_pat` and `build_expr` be
methods on `Model` that return `syn::Pat` and `syn::Expr` respectively,
to show a bit more type safety in the struct vs tuple variant handling.
As it stands though, the difference in the amount of code required is
extremely large, and not worth it.