Test fewer things in `bin/test`
The inclusion of those lines caused everything to need to be recompiled
if I ran `bin/test` a second time with no code changes. Ultimately this
script is just meant to be a quick "can I be semi-confident I didn't
break anything". It doesn't have to be fully comprehensive like CI.