Allow `now` to be used as a timestamptz
There are two things wrong with this PR, but I still think this is the
best short term solution.

The first and most glaringly obviously wrong thing about it is that the
type of `NOW()` **IS** `timestamptz`. However, since we currently have
that defined as a PG specific type, we can't set the type of the
expression properly. Timestamp with/without timezone is a thing in the
SQL standard, (it's just ignored by SQLite). I need to examine how MySQL
handles it and figure out how to promote it to a more general type.
However, I don't want to deal with that right now (and probably not
before 0.11). I do still want to fix this issue, so here we are.

We also have a more general problem of type equivalence that we need to
address. `Json` expressions can always be used where a `Jsonb` is
expected. `Text` with `VarChar` (though we've skirted that one by making
them type aliases). Again, this is a deeper problem that I don't
currently have a solution in mind for.

So in the short term, this gives us a way to punt on the problem for a
while. For the most common cases that crop up, we can write
`AsExpression` impls for those cases when we notice them.

Fixes #685.