Ensure aggregate functions enforce the column is from the right table
While working on #621, I noticed that these impls were incorrect and
could be used to compile an incorrect query. I've corrected the impls
and added the appropriate compile-fail test.

I'm not sure if this was just an oversight or if I intentionally did
this to avoid nullability somewhere. The latter is no longer relevant
since we always make these expressions nullable now.