Remove the `Clone` requirement from `StatementCache`
While there wasn't any performance concern from the reference counting,
SQLite and MySQL both use a struct that gets mutated as part of
execution. We want to ensure only one execution is occurring at a time,
and representing that with a unique reference better reflects (and
better enforces) that invariant.