Add chrono support for MySQL
Beyond the fact that the `ToSql` impl for `ffi::MYSQL_TIME` is
incredibly hacky, there's not a ton to say about this. I have no clue
what the fuck a negative date is, and the documentation doesn't state it
anywhere, so I'm just erroring if we get one. The range of dates that
MySQL accepts are quite small, so we're not worried about chrono
overflowing this time.