Ensure that `MysqlConnection` is `Send`
`mysql::Statement` is not send due to the raw pointer. The `MYSQL_STMT`
type is very much unsafe to be sent across threads, since it contains a
pointer to the connection which is not thread safe. However, we never
expose statements directly, and sending the whole connection is safe.

I've also just made it a supertrait of `Connection` so I don't make this
mistake again.

Fixes #726.