Release v0.11.3
This release addresses a memory safety issue in our interaction with
libmysqlclient. There was an undocumented assumption made by that
library which we were not taking into consideration, resulting in a
use-after-free bug. This was resulting in a segmentation fault on some
systems.

While this issue is quite serious, we are not issuing a CVE or a
security advisory. The bug resulted in 4 bytes being written to a
handful of values. It could not be exploited to cause a security issue.