Remove `.aliased` and `.with`
These features were poorly fleshed out. Their names imply that they work
for arbitrary aliases and use CTEs when in fact it's actually just a
cross join. The existence of these types will get in the way as we
revamp joins, and I think we can bring them back with a better API once
we've finished that.