Allow types from other crates with `table!`
This doesn't affect `infer_schema!` yet (which is where this is most
important), but it lays the groundwork required. Additional modules to
import types from can be specified by writing several `use` statements
before the table name. Due to how `use` works here, these will always
need to be relative to the crate root, and `self` or `super` will be
inconsistent.

This shouldn't matter, since the most common case is going to be to
import types from a third party crate, which means there will usually be
`extern crate whatever` at the root.

For the `table!` macro to finish, we now need to know all of the
following:

- table name
- schema name
- table body
- primary key
- type modules

Only two of those things are mandatory. I've refactored the macro to
just look for each missing optional piece and add it if it's not
present. The macro is a bit harder to read as a result, but it's better
than the "number of options factorial matchers" alternative.